+++ src/components/onboarding/bar/MainHeader.tsx
