+++ src/components/onboarding/first/MainFirst.tsx
