+++ src/components/onboarding/footer/MainFooter.tsx
