+++ src/components/onboarding/second/MainSecond.tsx
