+++ src/components/onboarding/third/MainThird.tsx
