+++ src/pages/MainPage.tsx
@@ -1,13 +1,13 @@
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import Header from '../components/Header';
-import MainFirst from '../components/main/MainFirst';
-import MainFooter from '../components/main/MainFooter';
-import MainFourth from '../components/main/MainFourth';
-import MainHeader from '../components/main/MainHeader';
-import MainSecond from '../components/main/MainSecond';
-import MainThird from '../components/main/MainThird';
-import StartNav from '../components/main/StartNav';
+import MainFirst from '../components/onboarding/MainFirst';
+import MainFooter from '../components/onboarding/MainFooter';
+import MainFourth from '../components/onboarding/MainFourth';
+import MainHeader from '../components/onboarding/MainHeader';
+import MainSecond from '../components/onboarding/MainSecond';
+import MainThird from '../components/onboarding/MainThird';
+import StartNav from '../components/onboarding/StartNav';
 
 const Container = styled.div`
   width: 100vw;
